pyproject.toml
[project]
name = "admin-portal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "awscli>=1.35.20",
    "awscli-plugin-endpoint>=0.4",
    "boto3>=1.35.54",
    "db-to-sqlite>=1.5",
    "django-anymail[mailgun]>=12.0",
    "django-countries>=7.6.1",
    "django-dirtyfields~=1.9.7",
    "django-dramatiq==0.10.0",
    "django-environ>=0.4.5",
    "django-extensions>=3.2.3",
    "django-model-utils>=5.0.0",
    "django-mysql>=4.15.0",
    "django-registration==3.1.2",
    "django-unixdatetimefield==1.0.3",
    "django[bcrypt]~=5.0.0",
    "djangorestframework-csv>=3.0.2",
    "djangorestframework>=3.15.2",
    "dramatiq[rabbitmq]>=1.17.1",
    "drf-yasg>=1.21.8",
    "gunicorn>=23.0.0",
    "mysqlclient==2.1.1",
    "phpserialize>=1.3",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "sentry-sdk>=2.17.0",
    "sqlalchemy~=1.4",
    "sqlite-utils>=3.37",
    "tld>=0.13",
    "uritemplate>=4.1.1",
    "whitenoise[brotli]~=4.1.4",
    "ipwhois>=1.3.0",
    "django-cors-headers>=4.6.0",
    "djangorestframework-jsonp>=1.0.2",
    "pillow>=11.0.0",
    "django-taggit-labels>=0.7.0",
    "django-taggit>=6.1.0",
    "django-waffle>=4.1.0",
    "django-tailwind>=3.8.0",
    "django-widget-tweaks>=1.5.0",
    "crispy-tailwind>=1.0.3",
    "django-crispy-forms>=2.3",
    "django-storages>=1.14.4",
    "django-autocomplete-light>=3.11.0",
    "markdown>=3.7",
    "toml>=0.10.2",
    "typing-extensions>=4.12.2",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "geoip2>=4.8.0",
    "iso3166>=2.1.1",
    "django-logentry-admin>=1.1.0",
    "rich>=13.9.4",
    "duckdb>=1.1.2",
    "dateutils>=0.6.12",
    "freezegun>=1.5.1",
    "django-filter>=24.3",
    "django-sql-explorer>=5.3",
    "django-admin-multiple-choice-list-filter>=0.1.1",
    "django-formtools>=2.5.1",
    "django-betterforms>=2.0.0",
    "django-convenient-formsets~=1.2.1",
    "django-basicauth>=0.5.3",
    "django-file-resubmit>=0.5.2",
    "django-guardian>=2.4.0",
    "drf-api-logger>=1.1.16",
    "urllib3>=2.2.3",
    "httpx>=0.28.1",
    "pytest-httpx>=0.35.0",
    "carbon-txt==0.0.18",
    "truststore>=0.10.4",
]

[tool.uv]
dev-dependencies = [
    "ansible-lint>=6.8.7",
    "black>=24.10.0",
    "coverage>=7.6.4",
    "django-browser-reload>=1.17.0",
    "django-debug-toolbar>=4.4.6",
    "django-upgrade>=1.22.1",
    "factory-boy>=3.3.1",
    "faker>=30.8.2",
    "flake8>=7.1.1",
    "furo>=2023.3.27",
    "hypothesis>=6.116.0",
    "ipdb>=0.13.13",
    "ipython>=8.29.0",
    "marimo>=0.9.14",
    "myst-parser>=1.0.0",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "pylint-django>=2.6.1",
    "pylint>=3.3.1",
    "pytest-cov>=6.0.0",
    "pytest-django>=4.9.0",
    "pytest-factoryboy>=2.7.0",
    "pytest-mock>=3.14.0",
    "pytest-xdist>=3.6.1",
    "pytest>=8.3.3",
    "sphinx-autobuild>=2024.10.3",
    "sphinx>=5.3.0",
    "sphinxcontrib-mermaid>=1.0.0",
]
